TICKET-4: [*] move validator to the services folder [+] lodash [*] refactoring validate function [-] unnecessary dependency
