+++ package.json
@@ -13,8 +13,8 @@
     "co-body": "^5.1.1",
     "koa": "~1.2.5",
     "koa-body": "~1.4.0",
-    "koa-bodyparser": "^2.5.0",
     "koa-request": "~1.0.0",
-    "koa-router": "~4.3.2"
+    "koa-router": "~4.3.2",
+    "lodash": "^4.17.4"
   }
 }
